#34 renamed and move to another package channel entity
