+++ slack-archive/build.gradle
@@ -49,6 +49,8 @@ dependencies {
     }
     compile 'org.springframework.boot:spring-boot-starter-jetty'
     compile 'org.springframework.boot:spring-boot-starter-data-mongodb'
+    compile 'org.apache.httpcomponents:httpclient:4.5.3'
+    compile group: 'org.apache.commons', name: 'commons-io', version: '1.3.2'
     compile 'org.projectlombok:lombok:1.16.10'
     compile 'javax.inject:javax.inject:1'
     testCompile 'net.javacrumbs.json-unit:json-unit-fluent:1.19.0'
