+++ slack-archive/src/main/resources/application.properties
@@ -4,3 +4,6 @@ spring.data.mongodb.uri=mongodb://root:root@localhost:27017/slack-archive
 spring.data.mongodb.messages.collection=message-archive
 spring.data.mongodb.channels.collection=channels
 spring.data.mongodb.repositories.enabled=true
+#Slack
+slack.api.token = xoxp-149859256610-149859256690-289297509942-300c51167bfc18618fd1fcef1e0c9037
+slack.api.channels.urltemplate = https://slack.com/api/channels.list?token=
